pyproject.toml
[tool.poetry]
name = "chemistry-be"
version = "0.1.0"
description = ""
authors = ["billy"]
readme = "README.md"
#packages = [{include = "chemistry_be"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.78.0"
uvicorn = ">=0.15.0,<0.16.0"
pydantic = ">=1.9.1,<2.0.0"
alembic = ">=1.8.1"
psycopg2-binary = ">=2.9.5"
pytest = ">=7.2.1"
httpx = ">=0.23.3"
pylint = "2.15.10"
isort = "5.11.4"
black = "22.12.0"
mypy = "0.991"
sqlalchemy-stubs = "0.4"
rdkit = "2023.3.3"
python-multipart = "0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
